Move DataframeView
